fix: correctness gaps in plan serialization deserialization paths
Five fixes from the PR #183 audit:

1. OpSerializationRegistry.RebuildMean: called ReduceSum and copied the
   sum to the output — so deserialized Mean ops produced the SUM of
   elements, off by a factor of N. Switched to ReduceMean (or TensorMean
   for the null-axes scalar case). This silently mis-served any plan
   that contained a Mean op.

2. OpSerializationRegistry.RebuildLeakyReLU: called eng.LeakyReLU(a,
   default!) — default(T) is 0, which collapses LeakyReLU to plain ReLU.
   Deserialized LeakyReLU plans silently lost their negative slope. Now
   reads alpha from savedState[0] (stored as double) and converts to T
   via MathHelper.GetNumericOperations<T>().FromDouble.

3. CompiledPlanLoader: the try/catch swallowed InvalidDataException from
   InferencePlanReader / TrainingPlanReader and returned null. Corruption,
   truncation, checksum mismatch, and wrong magic all collapsed into a
   silent cache miss — operators had no way to diagnose warm-start
   failures. Now logs the cause via Trace.WriteLine before returning
   null so the failure is visible but the contract is unchanged.

4. Plan format checksum typing: reader read storedChecksum as long,
   writer wrote (long)(ulong)checksum, and comparison was (long)computed
   != storedChecksum. The math works (same unchecked reinterpret on both
   sides) but drifts if a writer changes the cast. Reader now uses
   BitConverter.ToUInt64; writers write the raw ulong. Same bytes on
   disk, but reader/writer can no longer get out of sync.

5. ElementTypeCodeToName (both readers): dropped PlanFormatConstants
   .ElementTypeFloat16 into the "unknown-{code}" fallback, so any
   Float16 plan would fail compatibility even when the runtime could
   handle it. Added "System.Half" mapping; used a literal string instead
   of typeof(Half).FullName because System.Half doesn't exist on net471.

+++ src/AiDotNet.Tensors/Engines/Compilation/Serialization/InferencePlanWriter.cs
@@ -68,7 +68,7 @@ internal static class InferencePlanWriter
         stream.Write(bodyBytes, 0, bodyBytes.Length);
         using var footerWriter = new BinaryWriter(stream, Encoding.UTF8, leaveOpen: true);
         footerWriter.Write((long)bodyBytes.Length);
-        footerWriter.Write((long)checksum);
+        footerWriter.Write(checksum); // ulong — reader uses ToUInt64
     }
 
     private static void WriteStep<T>(BinaryWriter writer, CompiledStep<T> step, TensorIdMap<T> tensorMap)
